Remove python2 compatibility code
