+++ pyiron_base/generic/util.py
@@ -18,6 +18,22 @@ __status__ = "production"
 __date__ = "Sep 1, 2017"
 
 
+class Singleton(type):
+    """
+    Implemented with suggestions from
+
+    http://stackoverflow.com/questions/6760685/creating-a-singleton-in-python
+
+    """
+
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+
 def static_isinstance(obj, obj_type):
     """
     A static implementation of isinstance() - instead of comparing an object and a class, the object is compared to a
