+++ pyiron_base/job/script.py
@@ -2,7 +2,6 @@
 # Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
-from __future__ import print_function
 import os
 import shutil
 from pyiron_base.job.generic import GenericJob
