+++ pyiron_base/master/list.py
@@ -2,7 +2,6 @@
 # Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
-from __future__ import print_function
 from pyiron_base.generic.parameters import GenericParameters
 from pyiron_base.job.core import JobCore
 from pyiron_base.job.generic import GenericJob
