+++ pyiron_base/master/parallel.py
@@ -2,7 +2,6 @@
 # Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
-from __future__ import division, print_function
 from collections import OrderedDict
 from datetime import datetime
 import numpy as np
