+++ setup.py
@@ -40,7 +40,6 @@ setup(
         'psutil>=5.7.0',
         'pyfileindex>=0.0.4',
         'pysqa>=0.0.11',
-        'six>=1.14.0',
         'sqlalchemy>=1.3.14',
         'tables>=3.6.1'
     ],
